Small refactor to avoid looped imports.
