+++ src/dymad/utils/__init__.py
@@ -4,21 +4,26 @@ from dymad.utils.misc import load_config, setup_logging
 from dymad.utils.modules import ControlInterpolator, GNN, IdenCatGNN, IdenCatMLP, make_autoencoder, MLP, ResBlockGNN, ResBlockMLP
 from dymad.utils.plot import plot_summary, plot_trajectory, plot_hist
 from dymad.utils.prediction import predict_continuous, predict_discrete, predict_graph_continuous, predict_graph_discrete
+from dymad.utils.preprocessing import Compose, DelayEmbedder, Identity, make_transform, Scaler, SVD
 from dymad.utils.weak import generate_weak_weights
 from dymad.utils.sampling import TrajectorySampler
 from dymad.utils.scheduler import make_scheduler
 
 __all__ = [
+    "Compose",
     "ControlInterpolator",
+    "DelayEmbedder",
     "generate_weak_weights",
     "GNN",
     "IdenCatGNN",
     "IdenCatMLP",
+    "Identity",
     "load_checkpoint",
     "load_config",
     "load_model",
     "make_autoencoder",
     "make_scheduler",
+    "make_transform",
     "MLP",
     "plot_hist",
     "plot_summary",
@@ -30,7 +35,9 @@ __all__ = [
     "ResBlockGNN",
     "ResBlockMLP",
     "save_checkpoint",
+    "Scaler",
     "setup_logging",
+    "SVD",
     "TrajectorySampler",
-    "truncated_svd"
+    "truncated_svd",
 ]
