+++ tests/test_assert_trajmgr.py
@@ -2,7 +2,8 @@ import copy
 import numpy as np
 from pathlib import Path
 
-from dymad.data import DynData, make_transform, TrajectoryManager
+from dymad.data import DynData, TrajectoryManager
+from dymad.utils import make_transform
 
 HERE = Path(__file__).parent
