+++ tests/test_assert_trajmgr_graph.py
@@ -3,7 +3,8 @@ import numpy as np
 import torch
 from torch_geometric.utils import dense_to_sparse
 
-from dymad.data import DynGeoData, make_transform, TrajectoryManagerGraph
+from dymad.data import DynGeoData, TrajectoryManagerGraph
+from dymad.utils import make_transform
 
 def check_data(out, ref, label=''):
     for _s, _t in zip(out, ref):
